+++ rekos/adapters/__init__.py
@@ -3,14 +3,16 @@
 from .base import AdapterResult, BaseSourceAdapter, SourceRunResult
 from .maigret import MaigretAdapter
 from .sherlock import SherlockAdapter, SherlockUsernameAdapter
-from .web_osint import DnsDomainAdapter, EmailPassiveAdapter
+from .web_osint import DnsDomainAdapter, EmailEnrichmentAdapter, EmailPassiveAdapter, HibpBreachAdapter
 from .wmn import WmnUsernameAdapter
 
 __all__ = [
     "AdapterResult",
     "BaseSourceAdapter",
     "DnsDomainAdapter",
+    "EmailEnrichmentAdapter",
     "EmailPassiveAdapter",
+    "HibpBreachAdapter",
     "MaigretAdapter",
     "SherlockAdapter",
     "SherlockUsernameAdapter",